Eliminate a deprecation warning on Python 3.7
In Python 3.7, usage of collections.Sequence and collections.Mapping
produces a warning, as the canonical location is collections.abc.  This
commit eliminates the warning by conditionally importing the appropriate
module under the control of six.PY2.  This commit also bumps the version,
eliminates Python 3.4 as a supported version, and adds Python 3.7 as a
supported version.